Extract reloader into separate module
